install deb packages according to apt install dependency
To fix the following error:
--------
root@xps ~# dpkg -i gawk_1%3a4.1.4+dfsg-1_amd64.deb libreadline7_7.0-2_amd64.deb
dpkg: regarding gawk_1%3a4.1.4+dfsg-1_amd64.deb containing gawk, pre-dependency problem:          <======
 gawk pre-depends on libreadline7 (>= 6.0)
  libreadline7 is not installed.
--------
due to gawk*.deb's sort is in front of libreadline7*.deb, "gwak" pkg including "pre-dependency" definition.
the deb package installation sequence is necessary, and this sequence should be automatically generated rather than manually maintained.
so we adjust the install flow as:
   step 1 generate /opt/deb/sequence-deb by $PACKAGE_LIST and $LKP_SRC/distro/keep-deb
   step 2 generate /opt/deb/keep-deb by /opt/deb/sequence-deb
   step 3 tbox uses /opt/deb/keep-deb list to install deb from /opt/deb

+++ lib/bootstrap.sh
@@ -256,12 +256,27 @@ redirect_stdout_stderr()
 install_deb()
 {
 	local files
+	local filename
 
-	files="$(find /opt/deb -type f 2>/dev/null)" || return
-	[ -n "$files" ] || return
+	[ -d /opt/deb ] || return 0
 
-	dpkg -i $files || return
-	rm $files
+	if [ -s "/opt/deb/keep-deb" ]; then
+		while read -r filename
+		do
+			dpkg -i /opt/deb/$filename || {
+				echo "error: dpkg -i /opt/deb/$filename failed." 1>&2
+				return 1
+			}
+			rm /opt/deb/$filename
+		done < /opt/deb/keep-deb
+	# this part for backward-compatibility to handle old deps which doesn't have keep-deb
+	else
+		files="$(find /opt/deb -name '*.deb' -type f 2>/dev/null)"
+		[ -n "$files" ] || return
+
+		dpkg -i $files || return
+		rm $files
+	fi
 }
 
 fixup_packages()
